+++ src/main/java/Main.java
@@ -1,3 +1,5 @@
+import duke.Duke;
+
 import java.io.File;
 
 public class Main {
