+++ src/main/java/duke/tasklist/TaskList.java
@@ -1,15 +1,14 @@
-package tasklist;
+package duke.tasklist;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.ArrayList;
 
-import task.Task;
-import task.ToDo;
-import task.Deadline;
-import task.Event;
-
-import ui.Ui;
+import duke.task.Task;
+import duke.task.Event;
+import duke.task.ToDo;
+import duke.ui.Ui;
+import duke.task.Deadline;
 
 public class TaskList {
